// frontend/src/types.ts
export interface CustomInvoice {
  id?: string;
  company_id: string;
  company_name?: string;
  client_name?: string;
  description: string;
  amount: number;
  due_date: string;
  invoice_month?: string;
}

export interface InvoiceQuantity {
  quantity: number;
  price: number;
}